pass activity to factory
